Update BowlingCalc class
Added new system to calculate the final round. There are too many different scenarios that could happen if it is combined with the regular system. Separating it makes it simpler and easier to calculate.

also added a new getScore method for the bonus rounds
added the error checking because in next changes will be adding it to the rest of the code.